Moved checks to common_instrument_adapter
