+++ src/tests/unittests/configuration_helper/adapters/test_common_instrument_adapter.py
@@ -18,5 +18,5 @@ class TestCommonInstrumentAdapter(unittest.TestCase):
         config = PythonJsonStructure(name='new_name')
         adapter.apply(config)
         self.assertEqual('DummyInstrumentAdapter_dummy_address', adapter.name)
-        self.assertNotEqual('new_name', not(adapter.instrument.name))
+        self.assertNotEqual('new_name', adapter.instrument.name)
         adapter.close_instrument()
